Factor out init and set tests to own files
